addressing valgrind and ubsan issues [no ci]

+++ src/colour-from-r.c
@@ -15,13 +15,13 @@
 
 // Modified djb2 hash
  int djb2_hash(uint8_t *str) {
-     int hash = 5381;
+    unsigned int hash = 5381;
     int c;
 
     while ((c = *str++))
       hash = ((hash << 5) + hash) + c; /* hash * 33 + c */
 
-    return hash;
+    return (int)hash;
 }
